refactor(tools): extract scheduling model
