+++ packages/agent-core/benchmark/ToolScheduling.hs
@@ -16,13 +16,15 @@ import Agent.ToolDispatch
     ( functionToolCall
     , noArgsTool
     )
+import Agent.Tools.Scheduling
+    ( ToolAccess(..)
+    , ToolResource(..)
+    , ToolResourceClaim(..)
+    )
 import Agent.Tools.Types
     ( AppTool
     , ApprovalRule(..)
-    , ToolAccess(..)
     , ToolExecutionPolicy(..)
-    , ToolResource(..)
-    , ToolResourceClaim(..)
     , jsonAppToolWithExecution
     , mkToolRegistry
     , withToolResourceClaims
