+++ packages/agent-core/src/Agent/Tools/Types.hs
@@ -3,10 +3,6 @@ module Agent.Tools.Types
     , ToolSchema(..)
     , ApprovalRule(..)
     , ToolExecutionPolicy(..)
-    , ToolAccess(..)
-    , ToolResource(..)
-    , ToolResourceClaim(..)
-    , ToolSchedulingPlan(..)
     , ToolRegistry
     , ToolEnv(..)
     , defaultToolEnv
@@ -25,7 +21,6 @@ module Agent.Tools.Types
     , lookupRegisteredTool
     , toolExecutionPolicyFor
     , toolSchedulingPlanFor
-    , schedulingPlansConflict
     , dispatchRegisteredToolCall
     , jsonToolParameters
     , appToolHandlers
@@ -43,6 +38,12 @@ import Agent.ToolDispatch
     , dispatchToolHandler
     , handlerName
     )
+import Agent.Tools.Scheduling
+    ( ToolAccess(..)
+    , ToolResource(..)
+    , ToolResourceClaim(..)
+    , ToolSchedulingPlan(..)
+    )
 import Control.Exception.Safe (tryAny)
 import Control.Monad (foldM)
 import Data.Aeson (Value)
@@ -51,13 +52,6 @@ import qualified Data.Map.Strict as Map
 import Data.Text (Text)
 import qualified Data.Text as Text
 import System.OsPath (OsPath, dropTrailingPathSeparator)
-import System.OsPath
-    ( equalFilePath
-    , isAbsolute
-    , makeRelative
-    , splitDirectories
-    , unsafeEncodeUtf
-    )
 
 -- | Provider-facing schema. The sum prevents freeform tools from carrying
 -- meaningless JSON parameters.
@@ -81,29 +75,6 @@ data ApprovalRule
 data ToolExecutionPolicy
     = ParallelSafe
     | TurnSequential
-    deriving (Eq, Show)
-
-data ToolAccess
-    = ToolRead
-    | ToolWrite
-    deriving (Eq, Show)
-
-data ToolResource
-    = ToolAllPaths
-    | ToolPath !OsPath
-    | ToolPathTree !OsPath
-    | ToolNamedResource !Text
-    deriving (Eq, Show)
-
-data ToolResourceClaim = ToolResourceClaim
-    { claimAccess :: !ToolAccess
-    , claimResource :: !ToolResource
-    } deriving (Eq, Show)
-
-data ToolSchedulingPlan
-    = ToolUnconstrained
-    | ToolResourceClaims ![ToolResourceClaim]
-    | ToolExclusive
     deriving (Eq, Show)
 
 type ToolResourceResolver =
@@ -337,54 +308,6 @@ toolSchedulingPlanFor registry call =
                     ToolResourceClaims
                         [ToolResourceClaim ToolRead ToolAllPaths]
                 TurnSequential -> ToolExclusive
-
-schedulingPlansConflict
-    :: ToolSchedulingPlan
-    -> ToolSchedulingPlan
-    -> Bool
-schedulingPlansConflict ToolExclusive _ = True
-schedulingPlansConflict _ ToolExclusive = True
-schedulingPlansConflict ToolUnconstrained _ = False
-schedulingPlansConflict _ ToolUnconstrained = False
-schedulingPlansConflict (ToolResourceClaims left) (ToolResourceClaims right) =
-    or
-        [ claimsConflict leftClaim rightClaim
-        | leftClaim <- left
-        , rightClaim <- right
-        ]
-
-claimsConflict :: ToolResourceClaim -> ToolResourceClaim -> Bool
-claimsConflict left right =
-    resourcesOverlap left.claimResource right.claimResource
-        && (left.claimAccess == ToolWrite || right.claimAccess == ToolWrite)
-
-resourcesOverlap :: ToolResource -> ToolResource -> Bool
-resourcesOverlap ToolAllPaths ToolAllPaths = True
-resourcesOverlap ToolAllPaths ToolPath{} = True
-resourcesOverlap ToolAllPaths ToolPathTree{} = True
-resourcesOverlap ToolPath{} ToolAllPaths = True
-resourcesOverlap ToolPathTree{} ToolAllPaths = True
-resourcesOverlap (ToolNamedResource left) (ToolNamedResource right) =
-    left == right
-resourcesOverlap (ToolPath left) (ToolPath right) =
-    equalFilePath left right
-resourcesOverlap (ToolPathTree left) (ToolPath right) =
-    pathInside left right
-resourcesOverlap (ToolPath left) (ToolPathTree right) =
-    pathInside right left
-resourcesOverlap (ToolPathTree left) (ToolPathTree right) =
-    pathInside left right || pathInside right left
-resourcesOverlap _ _ = False
-
-pathInside :: OsPath -> OsPath -> Bool
-pathInside root path
-    | equalFilePath root path = True
-    | otherwise =
-        let relative = makeRelative root path
-        in not (isAbsolute relative)
-            && case splitDirectories relative of
-                first : _ -> first /= unsafeEncodeUtf ".."
-                [] -> True
 
 dispatchRegisteredToolCall
     :: ToolDispatchConfig
